requirements.txt
# Core OCR Libraries
easyocr==1.7.0
paddleocr==2.7.3
pytesseract==0.3.10

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0

# Web UI
streamlit==1.28.1
streamlit-file-uploader==0.0.1

# Database
sqlite3
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Visualization
matplotlib==3.7.2
plotly==5.17.0
